+++ src/structures/vroom/vehicle.h
@@ -61,7 +61,7 @@ struct Vehicle {
   bool has_same_profile(const Vehicle& other) const;
 
   Cost duration(Index i, Index j) const {
-    return cost_wrapper.cost(i, j);
+    return cost_wrapper.duration(i, j);
   }
 
   Cost cost(Index i, Index j) const {